Se añadió la parte grafica del analizador lexico

+++ Proyecto/prueba.java
@@ -1,6 +1,6 @@
 public class prueba {
     public static void main(String[] args) {
         System.out.println("Hola, Mundo");
-        float numero = 1111111111;
+        int numero = 1111111111;
     }
 }
